Rename interface files to match C++ convention + fix REBUILD sed
- face_menu.rs -> iface_menu.rs (CIFaceMenu)
- robot_builder.rs -> constructor.rs (CConstructor)
- Refresh CROSSREF.md interface table
- cargo fmt + clippy --all-targets: 0 warnings
- Fix REBUILD.md version-bump recipe: sed 's///e' corrupts the file
  (pipes whole pattern space to shell); use perl -i -pe '...e' instead
- Bump index.html cache-bust v=96 -> v=97

+++ rust_port/src/matrix_game/interface/history.rs
@@ -66,6 +66,7 @@ impl ConfigHistory {
     }
 
     /// Port of `CHistory::NextConfig` (CHistory.cpp:99-106).
+    #[allow(clippy::should_implement_trait)]
     pub fn next(&mut self) -> Option<RobotConfig> {
         if self.cursor < 0 || (self.cursor as usize) + 1 >= self.configs.len() {
             return None;
@@ -95,7 +96,7 @@ mod tests {
         let mut h = ConfigHistory::new();
         for i in 0..200 {
             let mut c = RobotConfig::new();
-            c.chassis.kind = RobotUnitKind((i + 1) as i32);
+            c.chassis.kind = RobotUnitKind(i + 1);
             h.add(c);
         }
         assert_eq!(h.configs.len(), 200);
